feat: implement conversation creation helper and update UI interaction tests
