+++ tests/desktop/helpers/conversationAgents/index.ts
@@ -1,4 +1,5 @@
 export * from "./conversationAgentTestData";
 export * from "./conversationAgentTestSuite";
+export * from "./conversationAgentUiHelpers";
 export * from "./setupConversationAgentTest";
 export * from "./setupMultipleConversationsWithAgents";
